+++ src/layman/map/__init__.py
@@ -51,11 +51,11 @@ PUBLICATION_TYPES = {
         # see also .util.TASKS_TO_MAP_INFO_KEYS
         'internal_sources': OrderedDict([
             ('layman.map.filesystem.uuid', InternalSourceTypeDef(info_items=[]),),
-            ('layman.map.prime_db_schema.table', InternalSourceTypeDef(info_items=['access_rights', 'name', 'title', 'uuid',
+            ('layman.map.prime_db_schema.table', InternalSourceTypeDef(info_items=['access_rights', 'name', 'title', 'description', 'uuid',
                                                                                    'bounding_box', 'native_crs', 'native_bounding_box',
                                                                                    'updated_at', 'type', 'id', 'image_mosaic',
                                                                                    'map_layers']),),
-            ('layman.map.filesystem.input_file', InternalSourceTypeDef(info_items=['description', 'file']),),
+            ('layman.map.filesystem.input_file', InternalSourceTypeDef(info_items=['file']),),
             ('layman.map.prime_db_schema.file_data', InternalSourceTypeDef(info_items=[]),),
             ('layman.map.filesystem.thumbnail', InternalSourceTypeDef(info_items=['thumbnail', ]),),
             ('layman.map.micka.soap', InternalSourceTypeDef(info_items=['metadata', ]),),